[Container] Move Provider class to abstract namespace.
