+++ src/Valkyrja/Api/Provider/ServiceProvider.php
@@ -18,7 +18,7 @@ use Valkyrja\Api\Manager\Api;
 use Valkyrja\Api\Manager\Contract\ApiContract;
 use Valkyrja\Application\Kernel\Contract\ApplicationContract;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
-use Valkyrja\Container\Provider\Provider;
+use Valkyrja\Container\Provider\Abstract\Provider;
 use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract;
 
 final class ServiceProvider extends Provider
