+++ src/Valkyrja/Cli/Interaction/Provider/ServiceProvider.php
@@ -19,7 +19,7 @@ use Valkyrja\Cli\Interaction\Data\Config;
 use Valkyrja\Cli\Interaction\Output\Factory\Contract\OutputFactoryContract;
 use Valkyrja\Cli\Interaction\Output\Factory\OutputFactory;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
-use Valkyrja\Container\Provider\Provider;
+use Valkyrja\Container\Provider\Abstract\Provider;
 
 final class ServiceProvider extends Provider
 {
