+++ src/Valkyrja/Container/Provider/Abstract/Provider.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Container\Provider;
+namespace Valkyrja\Container\Provider\Abstract;
 
 use Override;
 use Valkyrja\Container\Provider\Contract\ProviderContract;
