+++ src/Valkyrja/Sms/Provider/ServiceProvider.php
@@ -16,7 +16,7 @@ namespace Valkyrja\Sms\Provider;
 use Override;
 use Valkyrja\Application\Env\Env;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
-use Valkyrja\Container\Provider\Provider;
+use Valkyrja\Container\Provider\Abstract\Provider;
 use Valkyrja\Log\Logger\Contract\LoggerContract;
 use Valkyrja\Sms\Messenger\Contract\MessengerContract;
 use Valkyrja\Sms\Messenger\LogMessenger;
