+++ tests/Tests/Classes/Application/Provider/HttpRoutingDataProviderClass.php
@@ -15,7 +15,7 @@ namespace Valkyrja\Tests\Classes\Application\Provider;
 
 use Override;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
-use Valkyrja\Container\Provider\Provider;
+use Valkyrja\Container\Provider\Abstract\Provider;
 use Valkyrja\Http\Routing\Data\Data;
 
 final class HttpRoutingDataProviderClass extends Provider
