+++ tests/Tests/Unit/Cli/Routing/Generator/DataProviderFileGeneratorTest.php
@@ -63,11 +63,7 @@ final class DataProviderFileGeneratorTest extends TestCase
 
             namespace $namespace;
 
-            use Override;
-            use Valkyrja\Cli\Routing\Data\Data;
-            use Valkyrja\Container\Provider\Provider;
-            use Valkyrja\Container\Manager\Contract\ContainerContract;
-            use App\Provider\Data\CliRoutingData;
+            use App\Provider\Data\CliRoutingData;use Valkyrja\Cli\Routing\Data\Data;use Valkyrja\Container\Manager\Contract\ContainerContract;use Valkyrja\Container\Provider\Abstract\Provider;
 
             final class CliDataTestRoutingDataProvider extends Provider
             {
