+++ tests/Tests/Unit/Cli/Routing/Generator/DefaultDataProviderFileGeneratorTest.php
@@ -59,11 +59,7 @@ final class DefaultDataProviderFileGeneratorTest extends TestCase
 
             namespace $namespace;
 
-            use Override;
-            use Valkyrja\Cli\Routing\Data\Data;
-            use Valkyrja\Container\Provider\Provider;
-            use Valkyrja\Container\Manager\Contract\ContainerContract;
-            use Valkyrja\Cli\Routing\Provider\ServiceProvider;
+            use Valkyrja\Cli\Routing\Data\Data;use Valkyrja\Cli\Routing\Provider\ServiceProvider;use Valkyrja\Container\Manager\Contract\ContainerContract;use Valkyrja\Container\Provider\Abstract\Provider;
 
             final class CliRoutingDataTestDataProvider extends Provider
             {
