+++ tests/Tests/Unit/Container/Generator/DefaultDataProviderFileGeneratorTest.php
@@ -59,11 +59,7 @@ final class DefaultDataProviderFileGeneratorTest extends TestCase
 
             namespace $namespace;
 
-            use Override;
-            use Valkyrja\Container\Data\Data;
-            use Valkyrja\Container\Provider\Provider;
-            use Valkyrja\Container\Manager\Contract\ContainerContract;
-            use Valkyrja\Container\Provider\ServiceProvider;
+            use Valkyrja\Container\Data\Data;use Valkyrja\Container\Manager\Contract\ContainerContract;use Valkyrja\Container\Provider\Abstract\Provider;use Valkyrja\Container\Provider\ServiceProvider;
 
             final class ContainerDefaultDataTestDataProvider extends Provider
             {
