+++ tests/Tests/Unit/Http/Routing/Generator/DataProviderFileGeneratorTest.php
@@ -63,11 +63,7 @@ final class DataProviderFileGeneratorTest extends TestCase
 
             namespace $namespace;
 
-            use Override;
-            use Valkyrja\Http\Routing\Data\Data;
-            use Valkyrja\Container\Provider\Provider;
-            use Valkyrja\Container\Manager\Contract\ContainerContract;
-            use App\Provider\Data\HttpRoutingData;
+            use App\Provider\Data\HttpRoutingData;use Valkyrja\Container\Manager\Contract\ContainerContract;use Valkyrja\Container\Provider\Abstract\Provider;use Valkyrja\Http\Routing\Data\Data;
 
             final class HttpDataTestRoutingDataProvider extends Provider
             {
