+++ tests/Tests/Unit/Http/Routing/Generator/DefaultDataProviderFileGeneratorTest.php
@@ -59,11 +59,7 @@ final class DefaultDataProviderFileGeneratorTest extends TestCase
 
             namespace $namespace;
 
-            use Override;
-            use Valkyrja\Http\Routing\Data\Data;
-            use Valkyrja\Container\Provider\Provider;
-            use Valkyrja\Container\Manager\Contract\ContainerContract;
-            use Valkyrja\Http\Routing\Provider\ServiceProvider;
+            use Valkyrja\Container\Manager\Contract\ContainerContract;use Valkyrja\Container\Provider\Abstract\Provider;use Valkyrja\Http\Routing\Data\Data;use Valkyrja\Http\Routing\Provider\ServiceProvider;
 
             final class HttpRoutingDataTestDataProvider extends Provider
             {
